% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{assignSegs2Alts}
\alias{assignSegs2Alts}
\title{Assign segments to splicing alternatives}
\usage{
assignSegs2Alts(a, s)
}
\arguments{
\item{a}{alternatives, output of \code{\link{makeAlts}}}

\item{s}{segment annotation (first element of output of \code{\link{loadSAData}} function)}
}
\value{
segments with alt.id field
}
\description{
Cuts genes by constitutive segments into regions that
can be spliced alternatively
}
