% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{calcSAPvalue}
\alias{calcSAPvalue}
\title{Calculates p-values for GLM fits}
\usage{
calcSAPvalue(
  data.glm,
  overdisp = TRUE,
  disp.param = NULL,
  .parallel = FALSE,
  .progress = "none"
)
}
\arguments{
\item{data.glm}{list of glm fits from \code{\link{fitSAGLM}}}

\item{overdisp}{define how to deal with overdispersion:
if overdisp is TRUE (default), overdispersion will be taken into account (see Details),
if overdisp is FALSE binomial distribution will be used.}

\item{disp.param}{\itemize{
\item to use external disperison parameters
}}

\item{.parallel, }{.progress parameters for plyr::alply}
}
\value{
matrix with rows for each fit, firts column contains overdispersion, remainings contain p-values for each term
}
\description{
Calculate p-value based on chisq distribution for each fit and factor
Could account for overdispersion (see parameter qbinom)
}
\details{
if overdisp is TRUE, quasibinomial distribution will be used if overdispersion is more than 1,
binomial distibution will be used instead.
So, switch it off if you don't like to account for overdispersion or if you don't have replicates.
}
\examples{
alt.pv.bq = calcSAPvalue(alt.glm)

}
