% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{filterAlts}
\alias{filterAlts}
\title{Removes some kinds of segments from alternatives}
\usage{
filterAlts(a, filter.introns = F)
}
\arguments{
\item{a}{output of \code{\link{makeAlts}} method}

\item{filter.introns}{whether retained introns should be removed}
}
\value{
alternatives in the same format. Segments specified by parameters are removed. Alternatives without segments are removed from results.
}
\description{
Removes some kinds of segments from alternatives
}
