% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{fitDEGLM}
\alias{fitDEGLM}
\title{Fit binomial GLM for gene expression data
Reads from each sample considered as mapped within particular gene (success in binomial trials) and other (failure)}
\usage{
fitDEGLM(
  ge,
  formula,
  terms,
  pseudocount,
  .parallel = FALSE,
  .progress = "none"
)
}
\arguments{
\item{formula}{object returned by \code{\link{terms}} function. Should be in form of x ~ sum_of_factors}

\item{terms}{list with values of factors used in formula}

\item{pseudocount}{proportion of total (inclusion + exclusion) reads that will be added to both inclusion and exclusion counts to evide overfitting}

\item{.parallel, }{.progress parameters for plyr::alply}

\item{data}{object of sajr class, produced by \code{\link{loadGData}}}
}
\value{
list of glm objects
}
\description{
Fit binomial GLM for gene expression data
Reads from each sample considered as mapped within particular gene (success in binomial trials) and other (failure)
}
