% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{fitSAGLM}
\alias{fitSAGLM}
\title{Fit binomial GLM}
\usage{
fitSAGLM(
  data,
  formula,
  terms,
  pseudocount = 0,
  .parallel = FALSE,
  .progress = "none",
  return.pv = FALSE,
  overdisp = TRUE,
  disp.param = NULL
)
}
\arguments{
\item{data}{object of sajr class, produced by \code{\link{loadSAData}}}

\item{formula}{object returned by \code{\link{terms}} function. Should be in form of x ~ sum_of_factors}

\item{terms}{list with values of factors used in formula}

\item{pseudocount}{proportion of total (inclusion + exclusion) reads that will be added to both inclusion and exclusion counts to evide overfitting}

\item{.parallel, }{.progress parameters for plyr::alply}

\item{return.pv}{logical, should function return p-values instead of models. Makes calculations more memory-efficient.}

\item{overdisp}{should overdispersion be taked into account. Has effect only if return.pv is true, see \code{\link{calcSAPvalue}} for details}

\item{disp.param}{to use external disperison parameters. Has effect only if return.pv is true.}
}
\value{
list of glm objects
}
\description{
Fit GLM model with binomial distribution for each supplied segments
}
\examples{
mod = list(brain_reg=factor(c('pfc','pfc','cbc','cbc')),age=factor(c('nb','adult','nb','adult')))
alt.glm = fitSAGLM(seg.filtered,terms(x ~ brain_reg + age,keep.order=T),mod,0.05)

}
