% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{getPal}
\alias{getPal}
\title{Make sequential palette with n colors}
\usage{
getPal(col = c("blue", "white", "red"), n = 200)
}
\arguments{
\item{col}{colors to be used to construct palette}

\item{n}{desired number of colors}
}
\description{
Make sequential palette with n colors
}
