% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{loadIntrons}
\alias{loadIntrons}
\title{Loads intron annotations from gff file}
\usage{
loadIntrons(ann.gff)
}
\arguments{
\item{ann.gff}{name of annotation file}
}
\value{
intron_id,gene_id,chromosome,strand,start and stop
}
\description{
Loads intron annotations from gff file
}
