% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{mfi}
\alias{mfi}
\title{Finds all elements from within [x\link{1},x\link{2}]}
\usage{
mfi(x, v)
}
\arguments{
\item{x}{numeric vector with at least two elements, x\link{1} have to be not more than x\link{2}}

\item{y}{sorted numeric vector}
}
\value{
first and last indexes of elements in v that are within [x\link{1},x\link{2}]
}
\description{
Finds all elements from within [x\link{1},x\link{2}]
}
