% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{plotMDS}
\alias{plotMDS}
\title{Plots 2-dimensial MDS}
\usage{
plotMDS(
  d = NULL,
  dist.method = c("cor", "sp"),
  points = NULL,
  dist = NULL,
  pch = NULL,
  col = "black",
  main = "",
  text = NULL,
  ...
)
}
\arguments{
\item{d}{numerical matrix (row\emph{col=features}samples) of feature intensities. rows are segments, cols are samples}

\item{dist.method}{type of distance to be calculated. If first element is 'cor' than 1-cor distance is used with method specified as second element of dist.method parameter.
if first element is not 'cor' it is considered as method for dist function. For example dist.method='euclidean' or dist.method = c('cor','sp')}

\item{points}{points to plot. Will used instead of d and dist if specified.}

\item{dist}{distance matrix. Will used instead of d if specified}

\item{pch}{plot points instead of sample names}

\item{col}{graphical parameter}

\item{main}{graphical parameter}

\item{text}{sample names (colnames(d) used as default)}
}
\description{
Function plots 2-dimensial MDS for all supplied features
}
