% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sajr.R
\name{setSplSiteTypes}
\alias{setSplSiteTypes}
\title{Adds splicing site info to segments annotation}
\usage{
setSplSiteTypes(d, gff.fname)
}
\arguments{
\item{d}{sajr object (from loadSAData function)}

\item{gff.fname}{name of annotation file}
}
\value{
sajr object with additional info
}
\description{
Adds columns 'sites' to d$seg. Column contains information about splce sites
at segment boundaries: 'a' is acceptor, 'd' is donor, '.' means that both sites coincide.
's' and 'e' denotes TSS or polyA sites respectively
}
